requirements.txt
antlr4-python3-runtime==4.7.2
apt-clone==0.2.1
apturl==0.5.2
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.13.2
chardet==3.0.4
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
elasticsearch==7.0.4
elasticsearch-dsl==7.0.0
entrypoints==0.3
future==0.18.2
httplib2==0.9.2
idna==2.6
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kiwisolver==1.1.0
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==3.0.3
mistune==0.8.4
more-itertools==7.2.0
mrjob==0.7.0
nbconvert==5.6.1
nbformat==4.4.0
nemo-emblems==4.0.2
netifaces==0.10.4
networkx==2.3
notebook==6.0.1
numpy==1.16.3
onboard==1.4.1
PAM==0.4.2
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.0.0
psutil==5.4.2
ptyprocess==0.6.0
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
Pygments==2.4.2
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.0
pyRFC3339==1.0
pyrsistent==0.15.5
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
python-telegram-bot==12.2.0
python-xapp==1.4.0
python-xlib==0.20
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==18.1.0
qtconsole==4.5.5
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scipy==1.3.1
screen-resolution-extra==0.0.0
Send2Trash==1.5.0
sessioninstaller==0.0.0
setproctitle==1.1.10
six==1.11.0
system-service==0.3
systemd-python==234
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
xkit==0.0.0
zipp==0.6.0
